feat: collection view로 list 구현, segment 설정
