+++ OpenMarket/OpenMarket/Model/Product.swift
@@ -6,9 +6,9 @@ struct Product: Decodable, Hashable {
     let name: String
     let thumbnail: String
     let currency: Currency
-    let price: Double
-    let bargainPrice: Double
-    let discountedPrice: Double
+    let price: Int
+    let bargainPrice: Int
+    let discountedPrice: Int
     let stock: Int
     let createdAt: Date
     let issuedAt: Date
