+++ OpenMarket/OpenMarket/Model/ProductDetail.swift
@@ -12,9 +12,9 @@ struct ProductDetail: Decodable {
     let description: String
     let thumbnail: String
     let currency: Currency
-    let price: Double
-    let bargainPrice: Double
-    let discountedPrice: Double
+    let price: Int
+    let bargainPrice: Int
+    let discountedPrice: Int
     let stock: Int
     let images: [ProductDetailImage]
     let vendors: Vendor
